PP-3173 Read PaymentRequest when updating email
Change the code that passes round a ChargeEntity to use a
PaymentRequestEntity instead. This is because we are now using the new
PaymentRequest tables and want to start reading from them. Eventually
we will stop updating the Charges table and delete the code. Still need
to keep updating Charges until everything is updated but after this
change that should be trivial.

with @georgeracu